split policy tests into a file
